+++ src/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sleep Outside | Home</title>
-    <link rel="stylesheet" href="/css/style.css" />
+    <link rel="stylesheet" href="./css/style.css" />
 
     <script src="./js/main.js" type="module"></script>
   </head>
@@ -29,10 +29,23 @@
           </p>
         </div>
       </div>
-
-      <section class="products">
-        <h2>Top Products</h2>
-        <ul class="product-list"></ul>
+      <section class="categories">
+        <a href="product_listing/index.html?category=tents">
+          <img src="./public/icons/category-tents.svg" alt="Tents">
+          <span>Tents</span>
+        </a>
+        <a href="product_listing/index.html?category=backpacks">
+          <img src="./public/icons/category-backpacks.svg" alt="Backpacks">
+          <span>Backpacks</span>
+        </a>
+        <a href="product_listing/index.html?category=sleepingbags">
+          <img src="./public/icons/category-sleepingbags.svg" alt="Sleeping Bags">
+          <span>Sleeping Bags</span>
+        </a>
+        <a href="product_listing/index.html?category=hammocks">
+          <img src="./public/icons/category-hammocks.svg" alt="Hammocks">
+          <span>Hammocks</span>
+        </a>
       </section>
     </main>
     <footer id="footer"></footer>
