+++ src/product_listing/index.html
@@ -1,0 +1,22 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Sleep Outside | Product Listing</title>
+    <link rel="stylesheet" href="../css/style.css" />
+
+    <script src="../js/product-listing.js" type="module"></script>
+  </head>
+
+  <body>
+    <header class="divider" id="header"></header>
+    <main class="divider">
+      <section class="products">
+        <h2>Top Products</h2>
+        <ul class="product-list"></ul>
+      </section>
+    </main>
+    <footer id="footer"></footer>
+  </body>
+</html>
