+++ vite.config.js
@@ -12,6 +12,7 @@ export default defineConfig({
         cart: resolve(__dirname, 'src/cart/index.html'),
         checkout: resolve(__dirname, 'src/checkout/index.html'),
         product: resolve(__dirname, 'src/product_pages/index.html'),
+        productListing: resolve(__dirname, 'src/product_listing/index.html')
       },
     },
   },
